refactored BleManager.StateListener and BleManager.NativeStateListener to their own class files, updated all javadocs
